+++ sdk-lib/src/androidTest/java/cash/z/ecc/android/sdk/integration/SmokeTest.kt
@@ -4,11 +4,9 @@ import androidx.test.filters.LargeTest
 import androidx.test.filters.MediumTest
 import cash.z.ecc.android.sdk.annotation.MaintainedTest
 import cash.z.ecc.android.sdk.annotation.TestPurpose
-import cash.z.ecc.android.sdk.internal.service.LightWalletGrpcService
 import cash.z.ecc.android.sdk.util.TestWallet
 import kotlinx.coroutines.runBlocking
 import org.junit.Assert.assertEquals
-import org.junit.Assert.assertFalse
 import org.junit.Assert.assertTrue
 import org.junit.Ignore
 import org.junit.Test
@@ -20,16 +18,6 @@ import org.junit.Test
 @MaintainedTest(TestPurpose.COMMIT)
 @MediumTest
 class SmokeTest {
-
-    @Test
-    fun testNotPlaintext() {
-        val service =
-            wallet.synchronizer.processor.downloader.lightWalletService as LightWalletGrpcService
-        assertFalse(
-            "Wallet is using plaintext. This will cause problems for the test. Ensure that the `lightwalletd_allow_very_insecure_connections` resource value is false",
-            service.connectionInfo.usePlaintext
-        )
-    }
 
     @Test
     fun testFilePaths() {