+++ sdk-lib/src/androidTest/java/cash/z/ecc/android/sdk/integration/service/ChangeServiceTest.kt
@@ -14,8 +14,11 @@ import cash.z.ecc.android.sdk.internal.service.LightWalletGrpcService
 import cash.z.ecc.android.sdk.internal.service.LightWalletService
 import cash.z.ecc.android.sdk.internal.twig
 import cash.z.ecc.android.sdk.model.BlockHeight
+import cash.z.ecc.android.sdk.model.LightWalletEndpoint
+import cash.z.ecc.android.sdk.model.Mainnet
+import cash.z.ecc.android.sdk.model.Testnet
+import cash.z.ecc.android.sdk.model.ZcashNetwork
 import cash.z.ecc.android.sdk.test.ScopedTest
-import cash.z.ecc.android.sdk.type.ZcashNetwork
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.runBlocking
@@ -38,13 +41,15 @@ import org.mockito.Spy
 class ChangeServiceTest : ScopedTest() {
 
     val network = ZcashNetwork.Mainnet
+    val lightWalletEndpoint = LightWalletEndpoint.Mainnet
+    private val eccEndpoint = LightWalletEndpoint("lightwalletd.electriccoin.co", 9087, true)
 
     @Mock
     lateinit var mockBlockStore: CompactBlockStore
     var mockCloseable: AutoCloseable? = null
 
     @Spy
-    val service = LightWalletGrpcService(context, network)
+    val service = LightWalletGrpcService.new(context, lightWalletEndpoint)
 
     lateinit var downloader: CompactBlockDownloader
     lateinit var otherService: LightWalletService
@@ -53,7 +58,7 @@ class ChangeServiceTest : ScopedTest() {
     fun setup() {
         initMocks()
         downloader = CompactBlockDownloader(service, mockBlockStore)
-        otherService = LightWalletGrpcService(context, "lightwalletd.electriccoin.co")
+        otherService = LightWalletGrpcService.new(context, eccEndpoint)
     }
 
     @After
@@ -124,7 +129,7 @@ class ChangeServiceTest : ScopedTest() {
     fun testSwitchToInvalidServer() = runBlocking {
         var caughtException: Throwable? = null
 
-        downloader.changeService(LightWalletGrpcService(context, "invalid.lightwalletd")) {
+        downloader.changeService(LightWalletGrpcService.new(context, LightWalletEndpoint("invalid.lightwalletd", 9087, true))) {
             caughtException = it
         }
 
@@ -145,7 +150,7 @@ class ChangeServiceTest : ScopedTest() {
     fun testSwitchToTestnetFails() = runBlocking {
         var caughtException: Throwable? = null
 
-        downloader.changeService(LightWalletGrpcService(context, ZcashNetwork.Testnet)) {
+        downloader.changeService(LightWalletGrpcService.new(context, LightWalletEndpoint.Testnet)) {
             caughtException = it
         }