+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/internal/service/LightWalletGrpcService.kt
@@ -1,12 +1,10 @@
 package cash.z.ecc.android.sdk.internal.service
 
 import android.content.Context
-import cash.z.ecc.android.sdk.R
 import cash.z.ecc.android.sdk.annotation.OpenForTesting
-import cash.z.ecc.android.sdk.exception.LightWalletException
 import cash.z.ecc.android.sdk.internal.twig
 import cash.z.ecc.android.sdk.model.BlockHeight
-import cash.z.ecc.android.sdk.type.ZcashNetwork
+import cash.z.ecc.android.sdk.model.LightWalletEndpoint
 import cash.z.wallet.sdk.rpc.CompactFormats
 import cash.z.wallet.sdk.rpc.CompactTxStreamerGrpc
 import cash.z.wallet.sdk.rpc.Service
@@ -31,39 +29,14 @@ import kotlin.time.Duration.Companion.seconds
  */
 @OpenForTesting
 class LightWalletGrpcService private constructor(
+    context: Context,
+    private val lightWalletEndpoint: LightWalletEndpoint,
     var channel: ManagedChannel,
     private val singleRequestTimeout: Duration = 10.seconds,
     private val streamingRequestTimeout: Duration = 90.seconds
 ) : LightWalletService {
 
-    lateinit var connectionInfo: ConnectionInfo
-
-    constructor(
-        appContext: Context,
-        network: ZcashNetwork,
-        usePlaintext: Boolean =
-            appContext.resources.getBoolean(R.bool.lightwalletd_allow_very_insecure_connections)
-    ) : this(appContext, network.defaultHost, network.defaultPort, usePlaintext)
-
-    /**
-     * Construct an instance that corresponds to the given host and port.
-     *
-     * @param appContext the application context used to check whether TLS is required by this build
-     * flavor.
-     * @param host the host of the server to use.
-     * @param port the port of the server to use.
-     * @param usePlaintext whether to use TLS or plaintext for requests. Plaintext is dangerous so
-     * it requires jumping through a few more hoops.
-     */
-    constructor(
-        appContext: Context,
-        host: String,
-        port: Int = ZcashNetwork.Mainnet.defaultPort,
-        usePlaintext: Boolean =
-            appContext.resources.getBoolean(R.bool.lightwalletd_allow_very_insecure_connections)
-    ) : this(createDefaultChannel(appContext, host, port, usePlaintext)) {
-        connectionInfo = ConnectionInfo(appContext.applicationContext, host, port, usePlaintext)
-    }
+    private val applicationContext = context.applicationContext
 
     /* LightWalletService implementation */
 
@@ -141,22 +114,14 @@ class LightWalletGrpcService private constructor(
     }
 
     override fun reconnect() {
-        twig(
-            "closing existing channel and then reconnecting to ${connectionInfo.host}:" +
-                "${connectionInfo.port}?usePlaintext=${connectionInfo.usePlaintext}"
-        )
+        twig("closing existing channel and then reconnecting")
         channel.shutdown()
-        channel = createDefaultChannel(
-            connectionInfo.appContext,
-            connectionInfo.host,
-            connectionInfo.port,
-            connectionInfo.usePlaintext
-        )
+        channel = createDefaultChannel(applicationContext, lightWalletEndpoint)
     }
 
     // test code
-    var stateCount = 0
-    var state: ConnectivityState? = null
+    internal var stateCount = 0
+    internal var state: ConnectivityState? = null
     private fun requireChannel(): ManagedChannel {
         state = channel.getState(false).let { new ->
             if (state == new) stateCount++ else stateCount = 0
@@ -172,37 +137,13 @@ class LightWalletGrpcService private constructor(
         return channel
     }
 
-    //
-    // Utilities
-    //
-
-    private fun Channel.createStub(timeoutSec: Duration = 60.seconds) = CompactTxStreamerGrpc
-        .newBlockingStub(this)
-        .withDeadlineAfter(timeoutSec.inWholeSeconds, TimeUnit.SECONDS)
-
-    /**
-     * This function effectively parses streaming responses. Each call to next(), on the iterators
-     * returned from grpc, triggers a network call.
-     */
-    private fun <T> Iterator<T>.toList(): List<T> =
-        mutableListOf<T>().apply {
-            while (hasNext()) {
-                this@apply += next()
-            }
-        }
-
-    inner class ConnectionInfo(
-        val appContext: Context,
-        val host: String,
-        val port: Int,
-        val usePlaintext: Boolean
-    ) {
-        override fun toString(): String {
-            return "$host:$port?usePlaintext=$usePlaintext"
-        }
-    }
-
     companion object {
+        fun new(context: Context, lightWalletEndpoint: LightWalletEndpoint): LightWalletGrpcService {
+            val channel = createDefaultChannel(context, lightWalletEndpoint)
+
+            return LightWalletGrpcService(context, lightWalletEndpoint, channel)
+        }
+
         /**
          * Convenience function for creating the default channel to be used for all connections. It
          * is important that this channel can handle transitioning from WiFi to Cellular connections
@@ -210,30 +151,33 @@ class LightWalletGrpcService private constructor(
          */
         fun createDefaultChannel(
             appContext: Context,
-            host: String,
-            port: Int,
-            usePlaintext: Boolean
+            lightWalletEndpoint: LightWalletEndpoint
         ): ManagedChannel {
-            twig("Creating channel that will connect to $host:$port?usePlaintext=$usePlaintext")
+            twig(
+                "Creating channel that will connect to " +
+                    "${lightWalletEndpoint.host}:${lightWalletEndpoint.port}" +
+                    "/?usePlaintext=${!lightWalletEndpoint.isSecure}"
+            )
             return AndroidChannelBuilder
-                .forAddress(host, port)
+                .forAddress(lightWalletEndpoint.host, lightWalletEndpoint.port)
                 .context(appContext)
                 .enableFullStreamDecompression()
                 .apply {
-                    if (usePlaintext) {
-                        if (!appContext.resources.getBoolean(
-                                R.bool.lightwalletd_allow_very_insecure_connections
-                            )
-                        ) throw LightWalletException.InsecureConnection
-                        usePlaintext()
-                    } else {
+                    if (lightWalletEndpoint.isSecure) {
                         useTransportSecurity()
+                    } else {
+                        twig("WARNING: Using insecure channel")
+                        usePlaintext()
                     }
                 }
                 .build()
         }
     }
 }
+
+private fun Channel.createStub(timeoutSec: Duration = 60.seconds) = CompactTxStreamerGrpc
+    .newBlockingStub(this)
+    .withDeadlineAfter(timeoutSec.inWholeSeconds, TimeUnit.SECONDS)
 
 private fun BlockHeight.toBlockHeight(): Service.BlockID =
     Service.BlockID.newBuilder().setHeight(value).build()
@@ -243,3 +187,14 @@ private fun ClosedRange<BlockHeight>.toBlockRange(): Service.BlockRange =
         .setStart(start.toBlockHeight())
         .setEnd(endInclusive.toBlockHeight())
         .build()
+
+/**
+ * This function effectively parses streaming responses. Each call to next(), on the iterators
+ * returned from grpc, triggers a network call.
+ */
+private fun <T> Iterator<T>.toList(): List<T> =
+    mutableListOf<T>().apply {
+        while (hasNext()) {
+            this@apply += next()
+        }
+    }