+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/jni/RustBackendWelding.kt
@@ -4,8 +4,8 @@ import cash.z.ecc.android.sdk.internal.model.Checkpoint
 import cash.z.ecc.android.sdk.model.BlockHeight
 import cash.z.ecc.android.sdk.model.WalletBalance
 import cash.z.ecc.android.sdk.model.Zatoshi
+import cash.z.ecc.android.sdk.model.ZcashNetwork
 import cash.z.ecc.android.sdk.type.UnifiedViewingKey
-import cash.z.ecc.android.sdk.type.ZcashNetwork
 
 /**
  * Contract defining the exposed capabilities of the Rust backend.