+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/tool/CheckpointTool.kt
@@ -7,7 +7,7 @@ import cash.z.ecc.android.sdk.internal.from
 import cash.z.ecc.android.sdk.internal.model.Checkpoint
 import cash.z.ecc.android.sdk.internal.twig
 import cash.z.ecc.android.sdk.model.BlockHeight
-import cash.z.ecc.android.sdk.type.ZcashNetwork
+import cash.z.ecc.android.sdk.model.ZcashNetwork
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import java.io.BufferedReader